requirements/requirements.txt
numpy
pandas
sentence-transformers
docling
git+https://github.com/brandonstarxel/chunking_evaluation.git
beautifulsoup4
python-dotenv
requests
google-search-results
serpapi
snowflake-snowpark-python
boto3
streamlit
fastapi
uvicorn
langchain-pinecone
langchain-openai
langgraph
pygraphviz